Create Main.cpp
This is the Main.cpp file, with most of the included functions.  It reads values of the the Magnetometer, Accelerometer, Gyroscope, Altimeter, Ambient Light, Microphone Level, and prints all values, via USART, to a serial port. 

Includes the Master Write and Read functions for I2C communication, initialization functions for ADC, and several usefully data processing/debugging functions.