Create Printing.h
This is a file that is used exclusively for debugging purposes. The first function, debug_statement(), uses USART to communicate with a serial port on the PC, and can transmit either an int or string. The two ERROR() functions were used to debug the I2C communication protocol, and can be utilized however necessary.